requirements/default.txt
importlib
feedparser
celery
django-celery
BeautifulSoup
httplib2
html5lib
